requirements.txt
selenium~=4.4.2
PyYAML~=5.4.1
pytest~=7.1.1